requirements.txt
requests
pandas
python-binance
aiohttp
